+++ torchmetrics/image/lpip_similarity.py
@@ -39,10 +39,10 @@ def _valid_img(img: Tensor) -> bool:
 
 
 class LPIPS(Metric):
-    """The Learned Perceptual Image Patch Similarity (`LPIPS_`) is used to judge the perceptual similarity
-    between two images. LPIPS essentially computes the similarity between the activations of two image patches
-    for some pre-defined network. This measure have been shown to match human perseption well. A low LPIPS
-    score means that image patches are perceptual similar.
+    """The Learned Perceptual Image Patch Similarity (`LPIPS_`) is used to judge the perceptual similarity between
+    two images. LPIPS essentially computes the similarity between the activations of two image patches for some
+    pre-defined network. This measure have been shown to match human perseption well. A low LPIPS score means that
+    image patches are perceptual similar.
 
     Both input image patches are expected to have shape `[N, 3, H, W]` and be normalized to the [-1,1]
     range. The minimum size of `H, W` depends on the chosen backbone (see `net_type` arg).